Add tutorial video and button to Main.html
Introduced a 'Watch Tutorial' button and embedded a video element in Main.html to provide users with a visual guide. Updated styles to use rem units for better scalability and added logic to play the video in fullscreen on iOS devices. Also updated project files to include Ios-safari.mp4 as a resource.